+++ src/setup/tables/Mizar.Article.sql
