+++ src/setup/tables/Mizar.Theorem.sql
